Fix pump

+++ src/pump/PumpManager.cpp
@@ -28,7 +28,7 @@ void PumpManager::applySettings(bool enable, int cutoffRaw, int maxPower) {
         if (maxPower < 0) {
             maxPower = 0;
         }
-        maxPower_ = static_cast<uint8_t>(maxPower);
+        maxPower_ = maxPower;
     }
     bumpVersion();
 }
@@ -36,10 +36,10 @@ void PumpManager::applySettings(bool enable, int cutoffRaw, int maxPower) {
 bool PumpManager::isEnabled() const { return enable_; }
 int  PumpManager::getCutoffPressureRaw() const { return cutoffRaw_; }
 bool PumpManager::isPumpOn() const { return outputOn_; }
-uint8_t PumpManager::getMaxPower() const { return maxPower_; }
-uint8_t PumpManager::getCurrentPower() const { return currentPower_; }
+int PumpManager::getMaxPower() const { return maxPower_; }
+int PumpManager::getCurrentPower() const { return currentPower_; }
 
-void PumpManager::writePower_(uint8_t percent) {
+void PumpManager::writePower_(int percent) {
     if (percent > maxPower_) {
         percent = maxPower_;
     }
@@ -81,13 +81,13 @@ void PumpManager::evaluateAutoControl_() {
         return;
     }
 
-    int32_t errBeyond = difference - hysteresisRaw_;
-    int32_t range     = controlRangeRaw_;
+    int errBeyond = difference - hysteresisRaw_;
+    int range     = controlRangeRaw_;
     if (range < 1) {
         range = 1;
     }
 
-    int32_t target = (static_cast<int32_t>(maxPower_) * errBeyond) / range;
+    int target = (maxPower_ * errBeyond) / range;
 
     if (target > maxPower_) {
         target = maxPower_;
@@ -105,5 +105,5 @@ void PumpManager::evaluateAutoControl_() {
         target = minPower;
     }
 
-    writePower_(static_cast<uint8_t>(target));
+    writePower_(target);
 }
