+++ src/pump/PumpManager.h
@@ -19,11 +19,11 @@ public:
 
     int  getCutoffPressureRaw() const;
 
-    uint8_t  getMaxPower() const;
-    uint8_t  getCurrentPower() const;
+    int  getMaxPower() const;
+    int  getCurrentPower() const;
 
 private:
-    static constexpr uint8_t minPower = 10;
+    static constexpr int minPower = 10;
 
     IPumpAdapter& pumpAdapter_;
     PressureSensorManager& pressureSensor_;
@@ -34,10 +34,10 @@ private:
     int  hysteresisRaw_ = 100;
     int  controlRangeRaw_  = 800;
 
-    uint8_t  maxPower_      = 100;
-    uint8_t  currentPower_ = minPower;
+    int  maxPower_      = 100;
+    int  currentPower_ = minPower;
 
-    void writePower_(uint8_t percent);
+    void writePower_(int percent);
     void writeOff_();
     void evaluateAutoControl_();
 };
